Save memory by atomic put of smaller files
When not specifying the size of the uploaded object, minio
does multi-part upload with a partition size of 128 MiB.

Even when uploading an object of - say - 8 KiB in that way minio
drives RSS memory usage up to 600 MiB or so.

Whereas with atomic put it just uses 7 MiB or so.

Since the files are encrypted during upload it's not straight forward
to determine the result size a priori. This is because of possible
variable size openpgp headers and packet headers. Which might even
change between different golang-openpgp versions. Furthermore, when
enabling compression with openpgp encryption, the exact result size
simply can't be determined a priori.